Update client if requested
On IX sending updated client information on a document create request
does not update the client. To avoid doing it all the time, and also
to support a simple usage, we have a `:update-client` bool that if true
will make the update before creating the document.

Also changed the project name to include weareswat namespace (to
properly group it on clojars).